+++ example/scripts/build-ios.sh
@@ -4,10 +4,11 @@
 set -o pipefail
 set -e
 
-ARGS=$@
+XCBEAUTIFY_ARGS=$@
 
-xcodebuild -workspace ios/BitmovinPlayerReactNativeExample.xcworkspace \
+eval "xcodebuild -workspace ios/BitmovinPlayerReactNativeExample.xcworkspace \
     -scheme BitmovinPlayerReactNativeExample \
     -configuration Debug \
-    -destination 'generic/platform=iOS Simulator' \
-    -quiet | xcbeautify -qq --disable-logging $ARGS
+    -quiet \
+    ${XCODEBUILD_ARGS} \
+    | xcbeautify -qq --disable-logging $XCBEAUTIFY_ARGS"