+++ plugin/src/withBitmovinConfig.ts
@@ -3,13 +3,18 @@ import withBitmovinIosConfig from './withBitmovinIosConfig';
 import Features from './Features';
 import withBitmovinAndroidConfig from './withBitmovinAndroidConfig';
 
-const defaultFeatureFlags: Features = {
+const defaultFeatures: Features = {
   airPlay: false,
   backgroundPlayback: false,
   googleCastSDK: undefined,
   offline: false,
   pictureInPicture: false,
 };
+
+export interface BitmovinConfigOptions {
+  playerLicenseKey?: string;
+  features?: Features;
+}
 
 /**
  * Expo Config Plugin for Bitmovin Player.
@@ -18,7 +23,7 @@ const defaultFeatureFlags: Features = {
  * @param config - The Expo config object.
  * @param options - An object containing the player license key and feature flags.
  * @returns The modified Expo config object with Bitmovin Player configurations.
- * 
+ *
  * @example
  * // app.config.js
  * module.exports = {
@@ -27,7 +32,7 @@ const defaultFeatureFlags: Features = {
  *       'bitmovin-player-react-native',
  *       {
  *         playerLicenseKey: 'YOUR_BITMOVIN_PLAYER_LICENSE_KEY',
- *         featureFlags: {
+ *         features: {
  *           airPlay: true,
  *           backgroundPlayback: true,
  *           googleCastSDK: { android: '21.3.0', ios: '4.8.1.2' },
@@ -39,13 +44,11 @@ const defaultFeatureFlags: Features = {
  *   ]
  * };
  */
-const withBitmovinConfig: ConfigPlugin<{
-  playerLicenseKey: string;
-  featureFlags: Features;
-}> = (config, { playerLicenseKey, featureFlags }) => {
-  const features = { ...defaultFeatureFlags, ...(featureFlags || {}) };
-  config = withBitmovinIosConfig(config, { playerLicenseKey, features });
-  config = withBitmovinAndroidConfig(config, { playerLicenseKey, features });
+const withBitmovinConfig: ConfigPlugin<BitmovinConfigOptions> = (config, options) => {
+  const { playerLicenseKey, features } = options;
+  const mergedFeatures = { ...defaultFeatures, ...(features || {}) };
+  config = withBitmovinIosConfig(config, { playerLicenseKey, features: mergedFeatures });
+  config = withBitmovinAndroidConfig(config, { playerLicenseKey, features: mergedFeatures });
   return config;
 };